# specs/001-backup-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: BackupChrono API
  description: Central pull backup system with device-centric organization
  version: 1.0.0
  contact:
    name: BackupChrono
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://backup.example.com/api
    description: Production server
...
paths:
...
  /devices/{deviceId}/backups:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    
    get:
      tags: [devices, backups]
      summary: Get backup history for device
      operationId: getDeviceBackups
      responses:
        '200':
          description: Backup history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Backup'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== SHARES ====================
...
  /backups:
    get:
      tags: [backups]
      summary: List all backups
      operationId: listBackups
      parameters:
        - name: deviceId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by device ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of results
      responses:
        '200':
          description: List of backups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Backup'
  /backups/{backupId}:
    parameters:
      - $ref: '#/components/parameters/BackupId'
    
    get:
      tags: [backups]
      summary: Get backup details
      operationId: getBackup
      responses:
        '200':
          description: Backup details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /backups/{backupId}/logs:
    parameters:
      - $ref: '#/components/parameters/BackupId'
    
    get:
      tags: [backups]
      summary: Get backup execution logs
      operationId: getBackupLogs
      description: |
        Returns backup execution logs including warnings, errors, and detailed progress information.
        Basic backup statistics should be retrieved from GET /backups/{backupId}.
      responses:
        '200':
          description: Backup execution logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupLogs'
        '404':
          $ref: '#/components/responses/NotFound'
  /backups/{backupId}/files:
    parameters:
      - $ref: '#/components/parameters/BackupId'
      - name: path
        in: query
        schema:
          type: string
          default: /
        description: Directory path to browse
    
    get:
      tags: [backups]
      summary: Browse files in backup
      operationId: browseBackupFiles
      responses:
        '200':
          description: File listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileEntry'
        '404':
          $ref: '#/components/responses/NotFound'
  /backups/{backupId}/restore:
    parameters:
      - $ref: '#/components/parameters/BackupId'
    
    post:
      tags: [backups]
      summary: Restore files from backup
      operationId: restoreBackup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '202':
          description: Restore job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  restoreId:
                    type: string
                  status:
                    type: string
                    enum: [Pending, Running]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /backups/files/history:
    get:
      tags: [backups]
      summary: Get version history for a specific file across all backups
      operationId: getFileHistory
      parameters:
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: filePath
          in: query
          required: true
          schema:
            type: string
          description: Full path to file
      responses:
        '200':
          description: File version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileVersion'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ==================== CONFIGURATION ====================
...
components:
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BackupId:
      name: backupId
      in: path
      required: true
      schema:
        type: string
        description: Unique backup identifier
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Backup:
      type: object
      properties:
        id:
          type: string
          description: Unique backup identifier
        deviceId:
          type: string
          format: uuid
        shareId:
          type: string
          format: uuid
          description: Share ID for share-level backups (null for device-level)
        deviceName:
          type: string
        shareName:
          type: string
          description: Share name for share-level backups (null for device-level)
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [Success, Partial, Failed]
        sharesPaths:
          type: object
          description: Share names to paths mapping (all shares for device-level, single entry for share-level)
          additionalProperties:
            type: string
        fileStats:
          type: object
          description: File statistics
          properties:
            new:
              type: integer
              description: Number of new files
            changed:
              type: integer
              description: Number of changed files
            unmodified:
              type: integer
              description: Number of unmodified files
        dataStats:
          type: object
          description: Data processing statistics
          properties:
            added:
              type: integer
              format: int64
              description: Bytes of new data added
            processed:
              type: integer
              format: int64
              description: Total bytes processed
        duration:
          type: string
          description: ISO 8601 duration
        errorMessage:
          type: string
        createdByJobId:
          type: string
          format: uuid
          description: Backup job that created this backup
    BackupDetail:
      allOf:
        - $ref: '#/components/schemas/Backup'
        - type: object
          properties:
            directoryStats:
              type: object
              description: Directory statistics
              properties:
                new:
                  type: integer
                  description: Number of new directories
                changed:
                  type: integer
                  description: Number of changed directories
                unmodified:
                  type: integer
                  description: Number of unmodified directories
            snapshotInfo:
              type: object
              description: Snapshot metadata
              properties:
                snapshotId:
                  type: string
                  description: Unique identifier for this snapshot
                parentSnapshot:
                  type: string
                  description: Parent snapshot ID (for incremental backups)
                  nullable: true
                exitCode:
                  type: integer
                  description: Backup process exit code (0 = success)
            deduplicationInfo:
              type: object
              description: Deduplication and blob statistics
              properties:
                dataBlobs:
                  type: integer
                  description: Number of data blobs created
                treeBlobs:
                  type: integer
                  description: Number of tree blobs created
                ratio:
                  type: string
                  description: Deduplication ratio (e.g., "78.5%" or "2.87:1")
                spaceSaved:
                  type: string
                  description: Human-readable space saved by deduplication (e.g., "3.3 GB")
            shares:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  path:
                    type: string
                  fileCount:
                    type: integer
                  size:
                    type: integer
                    format: int64
    FileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        isDirectory:
          type: boolean
        size:
          type: integer
          format: int64
        modifiedAt:
          type: string
          format: date-time
        permissions:
          type: string
    FileVersion:
      type: object
      properties:
        backupId:
          type: string
          description: Backup ID containing this file version
        timestamp:
          type: string
          format: date-time
        size:
          type: integer
          format: int64
        hash:
          type: string
    RestoreRequest:
      type: object
      required: [targetPath]
      properties:
        targetPath:
          type: string
          description: Restore destination path
        includePaths:
          type: array
          items:
            type: string
          description: Specific files/folders to restore (empty = restore all)
        restoreToSource:
          type: boolean
          default: false
          description: Restore directly to original device/share location

    # ========== VALUE OBJECTS ==========
...
    BackupLogs:
      description: |
        Backup execution logs including warnings, errors, and progress information.
        All backup statistics and metadata should be retrieved from GET /backups/{backupId}.
      type: object
      properties:
        warnings:
          type: array
          items:
            type: string
          description: List of warning messages encountered during backup
          default: []
        errors:
          type: array
          items:
            type: string
          description: List of error messages encountered during backup
          default: []
        progressLog:
          type: array
          items:
            $ref: '#/components/schemas/ProgressLogEntry'
          description: Detailed progress log entries from backup execution
          default: []
    ProgressLogEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Log entry timestamp
        message:
          type: string
          description: Log message
        percentDone:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Backup completion percentage
        currentFiles:
          type: array
          items:
            type: string
          description: Files currently being processed
        filesDone:
          type: integer
          description: Number of files completed
        bytesDone:
          type: integer
          format: int64
          description: Number of bytes processed

    # ========== ERROR SCHEMA ==========
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true